+++ components/Layout.tsx
@@ -1,9 +1,13 @@
 import { MainHeader } from './MainHeader/MainHeader';
+import { MainNavigation } from './MainNavigation/MainNavigation';
 
 export const Layout: React.FC = ({ children }) => {
   return (
     <div>
-      <MainHeader />
+      <header>
+        <MainNavigation />
+        <MainHeader />
+      </header>
       <main>{children}</main>
       <footer></footer>
     </div>
