+++ components/MainHeader/mainHeader.module.scss
@@ -1,32 +1,11 @@
-.headerList {
-  display: flex;
-  list-style-type: none;
-  justify-content: center;
-  margin: 0;
-  padding: 1rem 0;
-  border-bottom: 1px solid var(--gray-border);
-
-  @media screen and (min-width: 45rem) {
-    padding: 2.2rem 0;
-  }
+.headerSection {
+  padding: 2rem;
+  box-shadow: 1px 3px 6px 1px var(--gray-shadow);
 }
 
-.headerItem {
-  margin: 0 1rem;
-  font-size: 1.25rem;
-  text-transform: uppercase;
-  font-weight: 700;
-}
-
-.label {
-  display: none;
-  @media screen and (min-width: 45rem) {
-    display: initial;
-  }
-}
-
-.icon {
-  margin-right: 5px;
-  line-height: 1.15;
-  vertical-align: bottom;
+.title {
+  width: 100%;
+  max-width: 500px;
+  padding: 0;
+  margin: 0 auto;
 }
